support linearized dynamics
The new LinearizedAgent class is an extension of NonlinearAgent in the
sense that it accepts a nonlinear state-update/output equations, but
linearizes the dynamics around a given trajectory.

You must call Planner.computeParameters(Xlin, Ulin) to update the
linearization before invoking Planner.optimize().

The linearization is done automatically in the simulator based on last
known predictions. To initialize the simulation, the user has to
provide initial linearization trajectory in simulator.UserData.Xlin and
simulator.UserData.Ulin.

See omp_demo9lin.m for an example.